Implementado chlan y inputlimit
- inputlimit: Limita el numero de caracteres que se pueden escribir en la consola
- chlan: Cambia el idioma de la shell
-- La carpeta lang contiene los diccionarios para los idiomas español e inglés
-- Ya están todos los prints adaptados al doble-idioma

+++ src/commands/cp.py
@@ -1,5 +1,7 @@
 import os
 from utils.log_gen import make_log
+from utils.curfew_utils import is_in_curfew
+from utils.lang_manager import t
 
 """
 - cp: Copia archivos
@@ -8,34 +10,39 @@ from utils.log_gen import make_log
 """
 def run(args : list[str]) -> None:
     if len(args) < 2: # Si no hay suficientes argumentos
-        print("Error: Se debe indicar el archivo de origen y el destino")
+        print(t("cp_missing_args"))
         make_log("cp", success=False, details="Argumentos insuficientes para el comando 'cp'")
         return
     elif len(args) > 2: # Si hay demasiados argumentos
-        print("Error: Demasiados argumentos para el comando 'cp'")
+        print(t("cp_too_many_args"))
         make_log("cp", success=False, details="Demasiados argumentos para el comando 'cp'")
         return
     
     src, dest = args[0], args[1]
+
+    if os.path.exists(dest) and is_in_curfew():
+        print(t("cp_restricted"))
+        make_log("cp", success=False, details="Comando restringido para sobreescribir durante el toque de queda") 
+        return
 
     # Si dest es un directorio existente, agregar el nombre del archivo
     if os.path.isdir(dest):
         dest = os.path.join(dest, os.path.basename(src))
     
     try:
-        with open(src, "rb") as f_src: # Abrir el archivo de origen en modo lectura binaria "rb" (debe ser binario para copiar cualquier tipo de archivo)
-            with open(dest, "wb") as f_dst: # Abrir/crear el archivo de destino en modo escritura binaria "wb"
+        with open(src, "rb") as f_src: # Abrir el archivo de origen en modo lectura binaria "rb"
+            with open(dest, "wb") as f_dst: # Abrir/crear archivo destino
                 while True:
-                    chunk = f_src.read(4096) # Leer el archivo de origen en bloques de 4096 bytes (Suele ser el tamaño comun)
+                    chunk = f_src.read(4096)
                     if not chunk:
                         break
                     f_dst.write(chunk)
         make_log(f"cp {src} {dest}", success=True, details=f"Archivo copiado exitosamente")
     except FileNotFoundError:
-        print(f"Error: El archivo '{src}' no existe")
+        print(t("cp_file_not_found").format(src))
         make_log(f"cp {src} {dest}", success=False, details=f"El archivo no existe")
         return
     except PermissionError:
-        print(f"Error: No tiene permiso")
+        print(t("cp_no_permission"))
         make_log(f"cp {src} {dest}", success=False, details="No tiene permiso")
         return
